Fix the way the upper radius is defined

+++ Radial Density Functions/new-bins_rad-density-func_uncertainties.py
@@ -139,7 +139,7 @@ def radial_density(partx, party, partz, mass, binsize, virrad, halox, haloy, hal
     virV = (4/3)*math.pi*(np.power((virrad+10),3)-np.power((virrad-10),3))
     virindex = np.where(np.logical_and(dis.astype(float)>float(virrad-10), dis.astype(float)<float(virrad+10)))[0]
     
-    mass = mass.astype(float)
+    #mass = mass.astype(float)
     virM = np.sum(mass[virindex])
     virdensity = virM/virV
